backend/requirements.txt
Django==6.0.4
djangorestframework==3.17.1
django-cors-headers==4.9.0
dj-database-url==3.0.1
gunicorn==23.0.0
psycopg[binary]==3.2.12
python-dotenv==1.2.2
whitenoise==6.11.0
